# Conflicts:
#	pom.xml
#	src/main/java/ro/societateahermes/backendservice/controller/UserControllerInterface.java
#	src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/UserController.java
#	src/main/java/ro/societateahermes/backendservice/entities/Event.java
#	src/main/java/ro/societateahermes/backendservice/service/EventServiceInterface.java
#	src/main/java/ro/societateahermes/backendservice/service/serviceImplementation/EventServiceImplementation.java
#	src/main/java/ro/societateahermes/backendservice/service/serviceImplementation/UserServiceImplementation.java
#	src/main/resources/application.properties

+++ src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/ImageControllerImplementation.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
+import java.util.List;
 
 
 import org.springframework.http.MediaType;
@@ -16,7 +17,10 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import ro.societateahermes.backendservice.controller.ImageControllerInterface;
 import ro.societateahermes.backendservice.entities.Image;
+import ro.societateahermes.backendservice.exceptions.UnathorizeException;
 import ro.societateahermes.backendservice.service.serviceImplementation.ImageServiceImplementation;
+import ro.societateahermes.backendservice.utils.PermissionChecker;
+import ro.societateahermes.backendservice.utils.RolesActiveUser;
 
 
 @RestController
@@ -31,8 +35,22 @@ public class ImageControllerImplementation implements ImageControllerInterface {
 
     @Override
     @PostMapping("/cd")
-    public String saveImage(@RequestParam("file") MultipartFile multipartFile) throws IOException {
+    public String saveImage(@RequestParam("file") MultipartFile multipartFile) throws IOException, UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.checkAdmin(roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         return imageService.convertMultiPartToFile(multipartFile, ImageType.CD);
+    }
+
+    @Override
+    @PostMapping("/{eventId}")
+    public String saveImageEvent(@PathVariable("eventId") long eventId, @RequestParam("file") MultipartFile multipartFile) throws IOException, UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.check(eventId, roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
+        return imageService.convertMultiPartToFile(multipartFile, ImageType.EVENTS);
     }
 
     @Override
@@ -44,8 +62,20 @@ public class ImageControllerImplementation implements ImageControllerInterface {
     }
 
     @Override
+    @GetMapping(produces = MediaType.APPLICATION_OCTET_STREAM_VALUE, value = "/events")
+    public ResponseEntity<?> getImageByPathForEvent(@RequestParam("name") String imageName) throws URISyntaxException, IOException {
+        File image = imageService.getImageByPath(imageName, ImageType.EVENTS);
+        byte[] imageContent = Files.readAllBytes(image.toPath());
+        return new ResponseEntity<>(imageContent, HttpStatus.OK);
+    }
+
+    @Override
     @DeleteMapping("/cd")
-    public void deleteImage(@RequestParam("name") String imageName) throws IOException, URISyntaxException {
+    public void deleteImage(@RequestParam("name") String imageName) throws IOException, URISyntaxException, UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.checkAdmin(roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         imageService.deleteImage(imageName, ImageType.CD);
     }
 }
